구매예약 TLog 추가
git-svn-id: https://skvalley.com/svn/repos/storeplatform/trunk/sac/storeplatform-sac@45213 01f5c84a-aead-4b29-b0f1-7c54dbeec05f

+++ src/main/java/com/skplanet/storeplatform/sac/purchase/order/controller/PurchaseOrderController.java
@@ -507,7 +507,7 @@ public class PurchaseOrderController {
 		final List<Long> prodPriceList = new ArrayList<Long>();
 		for (CreatePurchaseSacReqProduct product : req.getProductList()) {
 			prodIdList.add(product.getProdId());
-			prodPriceList.add((long) product.getProdAmt().doubleValue());
+			prodPriceList.add(product.getProdAmt().longValue());
 		}
 
 		new TLogUtil().set(new ShuttleSetter() {
